added prepareFAS
