+++ multi.fa
@@ -1,0 +1,12 @@
+>A0A1W5C924
+MSLNGRELVKQPPPADYNQSVLTMQKSVTITVTPQRSNSMDLLNFEEKRQLIASSLSLSDILHVSPAAAHVAKEVAASTVIATKKQNGTTMRSNSLNSGSRTPPLERKSSKFSAFGRLFKPWKWRRKKKSEKFESTSKSLERKISVRANREELVQKGILLPESPTSALSIADEEQYTIYNSVNGTNGAVLPNLATSVNAPNSTSIASNNVPLSQSAQQISSLGGGPGGVGTPVGSLNLGGGGGGGGSGASGSSGNGGPGGGGAGGAGVVLVSGLAHSQSAHQIGQPPPPSLVHAQLAPHQALQQQLQQHFANVAANASANMNQVQDNQRKTPPPANDGQQQQHQLQHHQMQQQTQSLQAQHHHQALQQQLQQHHQAQQQAQQLAQQHGQHGGGHGHTHGGHHGLPNHIVLSELPEPPIPVSEIGPIPPPPMFSTPSPTLAGTRSHGSSAGGVSMGGSSGGGGGGSSSSGGGGGIAGGGQPPNVHSTSSNSSIMSNLIPLDMNDYDYDDHENNDQEEIDSDEDYMYPMSQPNPNIDTNRIDEIPAKEPNINAVPLKSALKKKGSNPATPTQDNRALAVRYDTNPPAAATIKPVLRPRIRICSRPMRFQIGLPSTMENKENARPFVIRETGDDCDQSDGPIHYRDEDADRAAKIARKESLSIKLALRPDRQELINRNILQLQSDNEKQESKEAIGARLIRRLSMRPTAEELVERNILKTQSPAEEKKQKEEKKRYLLRKLSFRPTVDELKEKKIIRFNDYIEVTQAHDYDRRADKPWTRLTPKDKAAIRKELNEFKSSEMAVHEGSRHLTRFHRP
+>A0A1U7F4Y7
+MKSVQALLLLTLLAVAVMGDDDPDTKCGEKEVWDDCASSCQDICFEPPAELCDKKCNIGCFCEKGYVREYLDGGKCVRPDTYFRINCICGEEYDREYKDDEKCILPEHCIFTFRSKIGNLLA
+>A0A1S4GN82
+MLSSDIVGTIKSEKTQEICEGCGQKIKDRYFMKLSPDQFWHEQCLLCCICRIQLSQTCYTKNTKVYCKDDYYRIFGVSSVHQQNQQHHQQHGQQQQQQQHQRDCYGCGERIAPIEMVMRAKHLVYHLHCFLCYTCNRPLQKGEPFSIRAGKLICQHDLEKDFYGSAAAAAAAAAAAAAAAMHHHHGGAGAGLGPHLPPAAHLYGDGDDYLLDDGLRTRDGRRGPKRPRTILTSAQRRQFKASFDVSPKPCRKVREALAKDTGLSVRVVQVWFQNQRAKMKKILRKSKANANGATDGEKNHSDKEDKSIKLESPSSDHSHYLGLDSSYSSSSQPLNPNLPYSPDFPDNSDASLCSSDISLDESFDNVDEATSDTMSLQNLDLQPNLHSNGQLVNHNNNTIGSMMASNMAGSITHIDKLYLMQNSYFNSAQIEQ
+>A0A1S4GMD5
+MKKFRAFTCSALLLSALLAVTPSVVQARGIVSRRAFDSYRLPNTTVPTHYTLRLDTDIHREIFDYTGNVQIRINVLEATDQIVLHSSRSEITHVQLRNGVGLSVPVKNFEFDEDKEFLVVNVGTTLQPNTGTYTLEIDFTNSIARNDQAGFYRSSYDDENGVRRYLGLTQFESVDARTSFPCYDEPGIKTTYDILIVCGIDYHARSNAPLRGIQLLEGGKKLSTFATTPRMQTYLVAWLVSDFKYEREVLTDPQLAIASWSKPSSVHLLSYSVDASKRFMRAMEEYFDQLYSMSKIDNVAIKDEDYAAGAMENWGLVTYKESAISYGPEDGEQRQVGVVTIVGHEFTHQFFGNLLAPKWWSYLWLNEGFARLYQYYLGAISHPELQLRERFVTGPFQTALRADNSLTVRPMTYYTETRPSISRLFDSIAYDKSACVLRMMNYALGERTFQKGLRYYLQQNKERGVVEESNLFDSLEQAAKEDAVLPISLSIHEIFGSWSHQPGVPLVKVRRDGNEYLFTQERYIANTPPEEPFENSWWIPISFSTPSNNAFEKLPAFWMPPNVSEVAYTIETVQDEVVSFNPHSTGYYRVEYEESMLRELIERINKDHTGFEPATRARLIDDTLNIAQSRGGNYDVVLQLLNYLQRETDYVPWAVAYENLRHMQSMLRTNDAVSQLLERFIEKIASPLLERYGVDHRTGESASVGELRTIAIDLACSGSQSCNDDAQSAVRSLASKKHYTMEDNALLCHGLQQSNAEQRKALQAQFSQVLSDKNLDNPSAKQFLLQSLSCLGGDMTKVQSFAKYLLALSESDRTKALHRLLVELNVPLAEIVQSLQSELSQRRSNSVYYKHGPKLIFEIARYVTDPSDVQLLSSVASSYDSWVASEIEQQVNANLNWVEQNVEPLKNALTRYFDA
+>A0A1S4GN31
+MECCLSEEAKEQKRINQEIERQLRRDKRDARRELKLLLLGTGESGKSTFIKQMRIIHGSGYSDEDKRGFIKLVYQNIFMAMQSMIRAMDLLKILYSDPANIEHAELIRSVDFETVTTFEPPYVQAIKDLWADAGIQECYDRRREYQLTDSAKYYLSDLARIEAEDFLPTEQDILRARAPTTGILEYPFDLDSIIFRMVDVGGQRSERRKWIHCFENVTSIIFLVALSEYDQILFESENENRMEESKALFKTIITYPWFQHSSVILFLNKKDLLEEKIMYSHLVDYFPEYDGPQRDAIAAREFILRMFVDLNPDSEKIIYSHFTCATDTENIRFVFAAVKDTILQSNLKEYNLV
+>A0A1U7F5Z4
+MTNESKITPTELPGLHHQLPMLMPDGTLGNSSWTDLLNYTFENEALYHVNTSTIERIVSLMVPVFFGFIGLAGLLGNGLVVLVVAANQSMRSTTNLLIINLAAADLLFVVFCVPFTAVDYVLPEWPFGEYWCKFVQYMIVVTAHASAYTLVLMSLDRFLAVVHPITSMSIRTERNASLAICLTWFVILTTAIPVAISHGVLFYPRQGGYYTACLFLHNEGYSLVAFHRVTRMVVVVVLVFAFCWCPIQVILLLKSLKLYGLTHASIIFQIVSHVLAYTNSCINPVLYAFLSDNFRKAFRKVSKVVWCGKPPPPTIPPAAQTKTTRTTGNGTSNADAL
